+++ src/masque_h1_server.erl
@@ -114,32 +114,7 @@ build_dispatch(Opts) ->
         accept_bind => maps:get(accept_bind, Opts),
         resolver => maps:get(resolver, Opts, fun default_resolver/1),
         handler_opts => maps:merge(
-            maps:with(
-                [
-                    address_pool,
-                    routes,
-                    mtu,
-                    resolver,
-                    allow,
-                    family,
-                    allow_private,
-                    connect_timeout,
-                    socket_opts,
-                    bind_address,
-                    bind_port,
-                    bind_socket_opts,
-                    public_addresses,
-                    public_address_fun,
-                    peer_filter_fun,
-                    scrub_fun,
-                    allow_loopback,
-                    max_compression_contexts,
-                    max_compression_contexts_in,
-                    max_compression_contexts_out,
-                    max_pending_compression_responses
-                ],
-                Opts
-            ),
+            maps:with(masque_server:handler_opt_keys(), Opts),
             maps:get(handler_opts, Opts, #{})
         ),
         name => maps:get(drain_key, Opts, undefined)