+++ test/masque_auth_challenge_SUITE.erl
@@ -205,7 +205,7 @@ exercise_accept(Config) ->
     end.
 
 proxy_uri(Port) ->
-    iolist_to_binary(["https://localhost:", integer_to_list(Port)]).
+    iolist_to_binary(["https://127.0.0.1:", integer_to_list(Port)]).
 
 unique_name(Prefix) ->
     list_to_atom(